Add support for a custom highlighter function.
Existing code works fine when Prismjs is run in a browser to colorize code.
Howver, for static site builders such as Metalsmith, we would like to not
rely on browser Javascript. Setting an appopriate highlighter method on a
notebookjs object allows us to continue the rendering process using Prismjs
to highlight code and error cells.